Add check for KMS alias existence (#11)
**Problem**

AWS IoT RoboRunner offers the ability to automatically create a
shared space management priority queue if it doesn’t already exist
in the AWS account and there is a dedicated customer managed AWS
KMS key to use. By providing explicit verbage and checking for
the required AWS KMS key, we now make it easier to understand if the
key is missing.

**Solution**

We've implemented an explicit test for the sm-priority-queue-cmk
customer managed AWS KMS key at the start of operations. If the
key isn't found, a clear message will be logged to help isolate
the problem quickly.

**Testing**

Previous unit tests within PriorityQueueUtils have been updated to
reflect the changes made.

./gradlew clean build passes

Co-authored-by: Cameron Mine <camemine@amazon.com>