Add a utility to build a JSON for metric data
**Problem**

Variety of metrics will be emitted to Cloudwatch from the FMSG.
These metrics must be formatted as a JSON string so that they
can be processed by the Lambda function that publishes the
metrics to Cloudwatch. We will need a utility that allows
the creation of the data for each of these metrics easily without redundancy.

**Solution**

Create a utility method that will provide the functionality to create the metric data.

**Testing**

./gradlew clean build: `PASS` (includes all unit tests)
./gradlew clean test --tests com.amazon.iotroborunner.fmsg.utils.CloudWatchUtilsTest: `PASS`
Manually create a simple lambda and pass the output of this
method to lambda to see if it will publish to CloudWatch correctly: `PASS`.